requirements.txt
aiofiles==24.1.0
aiogram==3.16.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
appscript==1.3.0
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.0
docx2pdf==0.1.8
fire==0.7.0
fonttools==4.55.3
frozenlist==1.5.0
idna==3.10
lxml==5.3.0
magic-filter==1.0.12
multidict==6.1.0
numpy==2.2.1
opencv-python-headless==4.10.0.84
pandas==2.2.3
pdf2docx==0.5.8
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyMuPDF==1.25.1
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
reportlab==4.2.5
six==1.17.0
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
XlsxWriter==3.2.0
yarl==1.18.3
